pyproject.toml
[tool.poetry]
name = "belvo"
version = "0.1.0"
description = ""
authors = ["matfonseca <mfonseca@fi.uba.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.4"
pandas = "^2.3.1"
matplotlib = "^3.10.3"
langchain-openai = "^0.3.28"
tabulate = "^0.9.0"
langchain = "^0.3.27"
langgraph = "^0.6.2"
streamlit = "^1.47.1"
plotly = "^6.2.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"
ipykernel = "^6.30.0"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120
known_first_party = ["app"]
